missing

+++ backend/diet/src/main/java/fitrack/diet/repository/DietPlanRepository.java
@@ -4,6 +4,6 @@ package fitrack.diet.repository;
 import fitrack.diet.entity.DietPlan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DietPlanRepository extends JpaRepository<DietPlan, Integer> {
+public interface DietPlanRepository extends JpaRepository<DietPlan, Long> {
 
 }
